requirements.txt
django-cors-headers==3.13.0
django==4.1.3
djangorestframework==3.14.0
mysqlclient==2.1.1
pillow==9.3.0
pyjwt==1.7.1
redis==4.3.4
requests==2.28.1
